pyproject.toml
[project]
name = "protodune-metadex"
version = "0.1.0"
description = "ProtoDUNE Metadata Exchange (PROTODUNE-METADEX)"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "logfire>=3.25.0",
    "pydantic-settings>=2.10.1",
    "reflex>=0.8.2",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "mkdocs-git-authors-plugin>=0.10.0",
    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
    "mkdocs-glightbox>=0.4.0",
    "mkdocs-material>=9.6.15",
    "mkdocstrings-python>=1.16.12",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
    "ty>=0.0.1a15",
]
